REFACTOR split Combobox part from MedewerkerSearch component
