+++ src/features/search/MedewerkerSearch.vue
@@ -1,69 +1,22 @@
 <template>
-  <div
-    ref="divRef"
-    @keyup.down="nextIndex"
-    @keyup.up="previousIndex"
-    @keyup.enter="selectItem"
-  >
-    <input
-      v-bind="$attrs"
-      :id="inputId"
-      :required="required"
-      type="search"
-      placeholder="Zoek een medewerker"
-      autocomplete="off"
-      role="combobox"
-      :aria-expanded="showList ? 'true' : 'false'"
-      :aria-controls="listboxId"
-      :aria-owns="listboxId"
-      aria-autocomplete="list"
-      aria-haspopup="listbox"
-      v-model="searchText"
-      ref="inputRef"
-    />
-    <SimpleSpinner
-      v-if="!matchingResult && searchText && result.loading"
-      class="spinner"
-    />
-    <ul
-      v-if="showList"
-      class="utrecht-textbox"
-      role="listbox"
-      :id="listboxId"
-      :aria-labelledby="labelId"
-    >
-      <li
-        v-for="(r, i) in listItems"
-        :key="i"
-        @mouseover="handleHover(i)"
-        @mousedown="selectItem"
-        :class="{ active: i === activeIndex }"
-        role="option"
-      >
-        <article>
-          <header>{{ r.value }}</header>
-          <p>{{ r.description }}</p>
-        </article>
-      </li>
-    </ul>
-  </div>
+  <search-combobox
+    v-bind="{ ...$attrs, ...props }"
+    placeholder="Zoek een medewerker"
+    v-model="searchText"
+    :result="result"
+    :list-items="datalistItems"
+    :exact-match="true"
+    :required="true"
+  />
 </template>
 
-<script lang="ts">
-export default {
-  inheritAttrs: false,
-};
-</script>
-
 <script lang="ts" setup>
-import { computed } from "@vue/reactivity";
-import { debouncedRef, useFocusWithin } from "@vueuse/core";
-import { ref, watch } from "vue";
+import { debouncedRef } from "@vueuse/core";
+import { computed } from "vue";
 import { useGlobalSearch, useSources } from "./service";
 import type { SearchResult } from "./types";
-import SimpleSpinner from "../../components/SimpleSpinner.vue";
-import { nanoid } from "nanoid";
-import { focusNextFormItem } from "@/helpers/html";
+import SearchCombobox from "@/components/SearchCombobox.vue";
+import { mapServiceData } from "@/services";
 
 type DatalistItem = {
   value: string;
@@ -83,16 +36,7 @@ const props = defineProps({
     type: Boolean,
     default: false,
   },
-  defaultValue: {
-    type: String,
-    required: true,
-  },
 });
-
-const generatedLabelId = nanoid();
-const inputId = computed(() => props.id || generatedLabelId);
-const labelId = nanoid();
-const listboxId = nanoid();
 
 function mapDatalistItem(x: SearchResult): DatalistItem {
   const { contact, department, function: functie, user } = x?.jsonObject ?? {};
@@ -108,47 +52,15 @@ function mapDatalistItem(x: SearchResult): DatalistItem {
   };
 }
 
-const activeIndex = ref(0);
-
-function nextIndex() {
-  if (
-    showList.value &&
-    activeIndex.value < listItems.value.length - 1 &&
-    listItems.value.length
-  ) {
-    activeIndex.value += 1;
-  } else {
-    activeIndex.value = 0;
-  }
-  scrollIntoView();
-}
-
-function previousIndex() {
-  if (!showList.value || !listItems.value.length) {
-    activeIndex.value = 0;
-  } else if (activeIndex.value > 0) {
-    activeIndex.value -= 1;
-  } else {
-    activeIndex.value = listItems.value.length - 1;
-  }
-  scrollIntoView();
-}
-
-function selectItem() {
-  if (!listItems.value.length) return;
-  searchText.value = listItems.value[activeIndex.value].value;
-  focusNextFormItem(inputRef.value);
-}
-
 const emit = defineEmits(["update:modelValue"]);
 
-const inputRef = ref();
-const divRef = ref();
-
-const searchText = ref(props.defaultValue);
+const searchText = computed({
+  get: () => props.modelValue,
+  set(val) {
+    emit("update:modelValue", val);
+  },
+});
 const debouncedSearchText = debouncedRef(searchText, 300);
-
-const isScrolling = ref(false);
 
 const sources = useSources();
 
@@ -167,145 +79,8 @@ const searchParams = computed(() => {
   };
 });
 
-const hasFocus = useFocusWithin(divRef);
-
-const showList = computed(
-  () =>
-    listItems.value.length &&
-    hasFocus.focused.value &&
-    (!props.modelValue || props.modelValue !== searchText.value)
-);
-
 const result = useGlobalSearch(searchParams);
-
-const listItems = ref<DatalistItem[]>([]);
-
-watch(searchText, (t) => {
-  if (!t) {
-    listItems.value = [];
-  }
-});
-
-watch(result, (r) => {
-  if (r.error) {
-    listItems.value = [];
-    return;
-  }
-  if (!r.success || !r.data) return;
-  activeIndex.value = Math.max(
-    0,
-    Math.min(activeIndex.value, r.data.page.length - 1)
-  );
-  listItems.value = r.data.page.map(mapDatalistItem);
-});
-
-const matchingResult = computed(() => {
-  if (
-    props.modelValue === searchText.value ||
-    listItems.value.some((x) => x.value === searchText.value)
-  )
-    return searchText.value;
-  return "";
-});
-
-const shouldSetValidity = computed(
-  () => !matchingResult.value && props.required
+const datalistItems = mapServiceData(result, (paginated) =>
+  paginated.page.map(mapDatalistItem)
 );
-
-watch(matchingResult, (val) => {
-  emit("update:modelValue", val);
-});
-
-watch(
-  () => props.modelValue,
-  (v) => {
-    searchText.value = v;
-  }
-);
-
-watch([inputRef, shouldSetValidity, searchText], ([r, s, v]) => {
-  if (!(r instanceof HTMLInputElement)) return;
-  r.setCustomValidity(
-    s ? (!v ? "Zoek een medewerker" : "Kies een optie uit de lijst") : ""
-  );
-});
-
-function isInViewport(el: HTMLElement) {
-  const rect = el.getBoundingClientRect();
-  return (
-    rect.top >= 0 &&
-    rect.left >= 0 &&
-    rect.bottom <=
-      (window.innerHeight || document.documentElement.clientHeight) &&
-    rect.right <= (window.innerWidth || document.documentElement.clientWidth)
-  );
-}
-
-function handleHover(i: number) {
-  // ignore hover if we're scrolling, it's probably accidental
-  if (isScrolling.value) return;
-  activeIndex.value = i;
-}
-
-let timeoutId: number;
-
-function scrollIntoView() {
-  const el = divRef.value;
-  if (!(el instanceof HTMLElement)) return;
-  const matchingLi = el.getElementsByTagName("li").item(activeIndex.value);
-
-  if (matchingLi && !isInViewport(matchingLi)) {
-    // when we scroll, the cursor can accidentally end up on an item
-    // in that case, we don't want that item to be selected, because we are navigating using the arrow keys.
-    // so let's wait for a bit before allowing items to be selected on hover.
-    isScrolling.value = true;
-    matchingLi.scrollIntoView(false);
-    timeoutId && clearTimeout(timeoutId);
-    timeoutId = setTimeout(() => {
-      isScrolling.value = false;
-    }, 500);
-  }
-}
 </script>
-
-<style lang="scss" scoped>
-.spinner {
-  font-size: 0.875rem;
-}
-.spinner,
-input,
-ul {
-  grid-column: 1;
-  grid-row: 2;
-}
-div {
-  position: relative;
-  display: grid;
-  gap: var(--spacing-small);
-  align-items: center;
-}
-ul {
-  position: absolute;
-  border-radius: 0;
-  display: grid;
-  gap: var(--spacing-default);
-  align-self: flex-end;
-  transform: translateY(100%);
-}
-li {
-  max-width: 100%;
-  overflow-x: hidden;
-}
-
-li.active {
-  background-color: var(--color-secondary);
-  cursor: pointer;
-}
-article > p,
-article > header {
-  font-size: 0.875rem;
-}
-article > header {
-  font-weight: bold;
-}
-</style>
