+++ src/views/AfhandelingView.vue
@@ -290,7 +290,6 @@
                     >Contactverzoek versturen naar</label
                   >
                   <medewerker-search
-                    :defaultValue="vraag.contactverzoek.medewerker"
                     v-model="vraag.contactverzoek.medewerker"
                     :id="'verzoek-medewerker' + idx"
                     class="utrecht-textbox utrecht-textbox--html-input medewerker-search"
